+++ src/main/java/frc/robot/subsystems/shooter/ErectorIOSim.java
@@ -4,6 +4,7 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.wpilibj.simulation.SingleJointedArmSim;
 import frc.robot.constants.Constants;
+import frc.robot.subsystems.shooter.dynamics.ShooterState;
 
 public class ErectorIOSim implements ErectorIO {
   private final SingleJointedArmSim m_sim;
